+++ base/ClubParser.h
@@ -26,48 +26,6 @@ static const char* Address = "Address";
 static const char* LogoFile = "LogoFile";
 } // namespace Tags
 
-static Ipponboard::ClubList ParseJsonFile(QString filePath)
-{
-    Ipponboard::ClubList clubs;
-
-    QFile file(filePath);
-    if (!file.open(QIODevice::ReadOnly))
-    {
-        throw std::runtime_error(
-            QStringLiteral("Unable to open club configuration '%1'").arg(filePath).toStdString());
-    }
-
-    QJsonParseError parseError{};
-    const auto document = QJsonDocument::fromJson(file.readAll(), &parseError);
-    if (parseError.error != QJsonParseError::NoError)
-    {
-        throw std::runtime_error(QStringLiteral("Invalid club JSON '%1': %2")
-                                     .arg(filePath, parseError.errorString())
-                                     .toStdString());
-    }
-
-    if (!document.isArray())
-    {
-        throw std::runtime_error(
-            QStringLiteral("Expected JSON array in '%1'").arg(filePath).toStdString());
-    }
-
-    const auto jsonClubs = document.array();
-
-    for (auto const& jsonClub : jsonClubs)
-    {
-        const auto clubObject = jsonClub.toObject();
-        Ipponboard::Club club;
-        club.name = clubObject.value(QStringLiteral("name")).toString();
-        club.address = clubObject.value(QStringLiteral("address")).toString();
-        club.logoFile = clubObject.value(QStringLiteral("image")).toString();
-
-        clubs.push_back(club);
-    }
-
-    return clubs;
-}
-
 static Ipponboard::ClubList ParseIniFile(QString filePath)
 {
     QSettings settings(filePath, QSettings::IniFormat);